+++ src/main/java/woowacourse/exception/BusinessException.java
@@ -1,0 +1,12 @@
+package woowacourse.exception;
+
+public class BusinessException extends RuntimeException {
+
+	public BusinessException() {
+		super();
+	}
+
+	public BusinessException(String message) {
+		super(message);
+	}
+}
