+++ src/main/java/woowacourse/exception/InvalidCartItemException.java
@@ -1,6 +1,6 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.exception;
 
-public class InvalidCartItemException extends RuntimeException {
+public class InvalidCartItemException extends BusinessException {
     public InvalidCartItemException() {
         this("유효하지 않은 장바구니입니다.");
     }
