+++ src/main/java/woowacourse/exception/InvalidProductException.java
@@ -1,6 +1,6 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.exception;
 
-public class InvalidProductException extends RuntimeException {
+public class InvalidProductException extends BusinessException {
     public InvalidProductException() {
         this("올바르지 않은 사용자 이름이거나 상품 아이디 입니다.");
     }
