+++ src/main/java/woowacourse/shoppingcart/application/CartService.java
@@ -8,8 +8,8 @@ import woowacourse.shoppingcart.dao.CartItemDao;
 import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.exception.InvalidProductException;
-import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
+import woowacourse.exception.InvalidProductException;
+import woowacourse.exception.NotInCustomerCartItemException;
 
 import java.util.ArrayList;
 import java.util.List;
