--- src/main/java/woowacourse/shoppingcart/exception/InvalidCustomerException.java
@@ -1,11 +1,0 @@
-package woowacourse.shoppingcart.exception;
-
-public class InvalidCustomerException extends RuntimeException {
-    public InvalidCustomerException() {
-        this("존재하지 않는 유저입니다.");
-    }
-
-    public InvalidCustomerException(final String msg) {
-        super(msg);
-    }
-}
