+++ src/test/java/woowacourse/auth/application/AuthServiceTest.java
@@ -16,8 +16,8 @@ import woowacourse.auth.support.EncryptionStrategy;
 import woowacourse.auth.domain.Password;
 import woowacourse.auth.dto.token.TokenRequest;
 import woowacourse.auth.dto.token.TokenResponse;
-import woowacourse.auth.exception.InvalidAuthException;
-import woowacourse.auth.exception.InvalidCustomerException;
+import woowacourse.exception.InvalidAuthException;
+import woowacourse.exception.InvalidCustomerException;
 import woowacourse.auth.support.JwtTokenProvider;
 
 @ExtendWith(MockitoExtension.class)
