+++ src/test/java/woowacourse/auth/application/CustomerServiceTest.java
@@ -20,8 +20,8 @@ import woowacourse.auth.domain.Password;
 import woowacourse.auth.dto.customer.CustomerDeleteRequest;
 import woowacourse.auth.dto.customer.CustomerRequest;
 import woowacourse.auth.dto.customer.CustomerUpdateRequest;
-import woowacourse.auth.exception.InvalidAuthException;
-import woowacourse.auth.exception.InvalidCustomerException;
+import woowacourse.exception.InvalidAuthException;
+import woowacourse.exception.InvalidCustomerException;
 import woowacourse.auth.support.EncryptionStrategy;
 
 @ExtendWith(MockitoExtension.class)
